requirements.txt
httpx==0.27.2
imap-tools==1.7.2
loguru==0.7.2
playwright==1.46.0
tenacity==9.0.0
tortoise-orm==0.21.6
